+++ examples/rustapi_module/src/othermod.rs
@@ -30,7 +30,7 @@ fn double(x: i32) -> i32 {
 
 #[pymodinit]
 fn othermod(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_function!(double))?;
+    m.add_wrapped(wrap_function!(double))?;
     m.add_class::<ModClass>()?;
 
     m.add("USIZE_MIN", usize::min_value())?;