+++ examples/word-count/src/lib.rs
@@ -80,7 +80,7 @@ fn count_line(line: &str, needle: &str) -> usize {
 
 #[pymodinit]
 fn word_count(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_function!(count_line))?;
+    m.add_wrapped(wrap_function!(count_line))?;
     m.add_class::<WordCounter>()?;
 
     Ok(())